Refactoring - move LoadingView and NoResultsView onto SearchResultsViewController - extract BookCell into its own file - add helper extension to center UIViews - move handle() and prepareForSearch() into SearchResultsViewController
